requirements.txt
pandas==2.2.0
numpy==1.26.3
scipy==1.12.0
statsmodels==0.14.1
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.2
openpyxl==3.1.2
jupyter==1.0.0
notebook==7.0.7
ipykernel==6.29.0
plotly==5.18.0
python-pptx==0.6.23